+++ packages/firebase_storage/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+* Added StorageReference delete function to remove files from Firebase.
+
 ## 0.1.1
 
 * Simplified and upgraded Android project template to Android SDK 27.